requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.7.0
Flask-Login==0.5.0
Flask-WTF==0.14.3
flask-mail==0.9.1
email-validator==1.1.2
SQLAlchemy==1.3.0
pyjwt==2.1.0